#styleProject In this commit I create the Products components, I organized the code and I changed the name of some variables
