+++ lib/screens/produtos.dart
@@ -1,4 +1,5 @@
 import 'dart:ui';
+import 'package:App/components/CustomProducts.dart';
 import 'package:App/service/api.dart';
 import 'package:App/style.dart';
 import 'package:flutter/material.dart';
@@ -26,50 +27,6 @@ class _ProdutosPageState extends State<ProdutosPage> {
   void initState() {
     getData();
     super.initState();
-  }
-
-  List<Widget> indexItems(List<dynamic> items) {
-    List<Widget> itemsWidget = [];
-
-    for (int i = 0; i < items.length; i++) {
-      itemsWidget.add(Container(
-        padding: EdgeInsets.only(top: 20),
-        alignment: Alignment.center,
-        width: 150,
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.start,
-          children: [
-            Padding(
-              padding: const EdgeInsets.only(
-                left: 5,
-                right: 5,
-              ),
-              child: Image.network(
-                'https://assets.instabuy.com.br/ib.item.image.small/s-${items[i]['images'][0]}',
-              ),
-            ),
-            Column(
-              crossAxisAlignment: CrossAxisAlignment.center,
-              children: [
-                Container(
-                  padding: EdgeInsets.all(5),
-                  child: Text(
-                    '${items[i]['name']}',
-                    textAlign: TextAlign.center,
-                  ),
-                ),
-                Container(
-                  padding: EdgeInsets.all(5),
-                  child: Text('R\$' +
-                      '${items[i]['prices'][0]['price'].toStringAsFixed(2)}'),
-                )
-              ],
-            ),
-          ],
-        ),
-      ));
-    }
-    return itemsWidget;
   }
 
   @override
