+++ lib/style.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
 
-const Color colorInstaB = Color(0xFFFF7E00);
+const Color orangeInstaB = Color(0xFFFF7E00);
+const Color greenInstaB = Color(0xff008004);
 const Color backGroundApp = Color(0xFFDBA5);
